+++ example/play.ts
@@ -1,13 +1,11 @@
 import { resolve } from 'path';
 import { generateDeclaration } from '../src';
 
-console.log(
-  generateDeclaration({
-    name: 'piral-sample',
-    root: resolve(__dirname, '..', 'tests', 'assets'),
-    files: ['doc1.ts'],
-    types: ['doc1.ts'],
-    apis: [],
-    imports: [],
-  }),
-);
+generateDeclaration({
+  name: 'piral-sample',
+  root: resolve(__dirname, '..', 'tests', 'assets'),
+  files: ['prop1.ts'],
+  types: ['prop1.ts'],
+  apis: [],
+  imports: [],
+}).then((res) => console.log(res));