+++ tests/assets/symbol1.ts
@@ -1,0 +1,3 @@
+export interface Foo {
+  readonly hasInstance: unique symbol;
+}