+++ tests/assets/symbol2.ts
@@ -1,0 +1,3 @@
+export interface String {
+    split(splitter: { [Symbol.split](string: string, limit?: number): string[]; }, limit?: number): string[];
+}